+++ src/field_region_map.c
@@ -42,7 +42,7 @@ static EWRAM_DATA struct {
     MainCallback callback;
     u32 unused;
     struct RegionMap regionMap;
-    u16 state;
+    u32 state;
 } *sFieldRegionMapHandler = NULL;
 
 static void MCB2_InitRegionMapRegisters(void);
